refactor: create a common function
